Add support for UEFI
After the conversion to RHEL system on UEFI, usually these errors are
present:
 1) Missing grub files (grubenv, grub.cfg, ..) insinde the
    /boot/efi/EFI/redhat directory [CentOS Linux only]
 2) Current (default) bootloader points to non-existing files
    [CentOS Linux only]
 3) New bootloader entry for RHEL doesn't exist
    (in case the original entry points to the efi redhat directory,
     just the label is missleading, speaking about the orig OS)

For CentOS, the first two issues are critical and lead into the broken
bootloader. For OL this is not blocking as the redhat efi dir is used
already (paths are same by default). This commit is resolving these
problems, as it enshurs that all expected files are present inside
the redhat efi dir and creates a new EFI bootloader entry pointing
to the default EFI binary file:
  \\EFI\\redhat\\shimx64.bin

TODO:
 - [ ] use grubx64.efi in case the shim package is not installed
       (or handle the problem in a different way)
 - [ ] What to do with other ORIG_OS entries EFI bootloaders?
       Additional centos / OL bootloaders could contain missleading
       labels speaking about original OS which is not present anymore.
       However, if points e.g. to an LVM snapshot with centos..?
       Do we need to deal with this?
 - [ ] Remove the original efi centos directory when no other efi
       bins are present
 - [ ] How to deal with errors? Print warnings? Critical?...
 - [ ] What about systems using completely different directory for
       the default bootloader?