UEFI: grub library: refactorring + extend EFIBootLoader
Added the EFI_MOUNTPOINT constant to refer to the expected /boot/efi
mountpoint. All related paths in the grub module are updated to use
the contant.

As a preparation for additional checks, the EFIBootLoader class is
extended by new methods: get_canonical_path, is_referring_to_file.
These methods inform user whether an entry points to an EFI binary
file and what is the location of the file.