feat: verify and prevent double rhel efi entries
Current implementation was just adding a new entry without checking if a
RHEL entry was already present. This adds a check to make sure we are
not duplicating the entries.

A check is also implemented for removing the old EFI entry, to make sure
the original entry exists before checking if it's modified. This should
prevent any index out of bounds issues.